ZJIT: Convert inliner tests without a dependency on external libraries into codegen tests
