+++ test/ruby/test_zjit.rb
@@ -382,258 +382,6 @@ class TestZJIT < Test::Unit::TestCase
     }, call_threshold: 14, num_profiles: 5
   end
 
-  def test_inlined_method_returns_correct_value
-    assert_runs '3', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def add_one(x) = x + 1
-      def test(n) = add_one(n)
-
-      test(2)
-    RUBY
-  end
-
-  def test_inlined_method_deoptimizes_on_redefinition
-    assert_runs '100', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def callee(x) = x + 1
-      def test(n) = callee(n)
-
-      # Warm up so callee gets inlined into test.
-      test(1)
-      test(1)
-
-      # Redefine callee. The inlined version in test should be invalidated,
-      # and subsequent calls should use the new definition.
-      def callee(x) = x * 100
-
-      test(1)
-    RUBY
-  end
-
-  def test_inlined_method_survives_compact_between_calls
-    omit("GC.compact support is required for this test") unless GC.respond_to?(:compact)
-    assert_runs '2', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def callee(x) = x + 1
-      def test(n) = callee(n)
-
-      # Warm up so callee gets inlined into test.
-      test(1)
-      test(1)
-
-      # Compaction moves Ruby objects, including any VALUEs the inlined body
-      # references through Const::Value (e.g. the Integer literal 1 in callee).
-      # The payload's GC-offset tracking must update those references in place.
-      GC.compact
-
-      test(1)
-    RUBY
-  end
-
-  def test_inlined_method_survives_compact_during_call
-    omit("GC.compact support is required for this test") unless GC.respond_to?(:compact)
-    assert_runs '2', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def trigger_compact = GC.compact
-      def callee(x)
-        trigger_compact
-        x + 1
-      end
-      def test(n) = callee(n)
-
-      # Warm up so callee gets inlined into test.
-      test(1)
-      test(1)
-
-      # trigger_compact fires compaction while the inlined body is live on the
-      # stack. The JITFrame attached to cfp->jit_return for the inlined frame
-      # holds the callee's ISEQ pointer, which compaction will move. Without
-      # the rb_zjit_jit_frame_update_references hook, the on-stack JITFrame
-      # would point at a moved ISEQ and subsequent frame inspection (backtrace,
-      # materialization on side exit) would crash.
-      test(1)
-    RUBY
-  end
-
-  def test_inlined_method_with_required_keyword
-    assert_runs '12', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def callee(x, y:) = x + y
-      def test(n) = callee(n, y: 10)
-
-      test(2)
-    RUBY
-  end
-
-  def test_inlined_method_with_optional_keyword_supplied
-    assert_runs '7', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def callee(x, y: 100) = x + y
-      def test(n) = callee(n, y: 5)
-
-      test(2)
-    RUBY
-  end
-
-  def test_inlined_method_with_optional_keyword_omitted
-    assert_runs '102', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def callee(x, y: 100) = x + y
-      def test(n) = callee(n)
-
-      test(2)
-    RUBY
-  end
-
-  def test_inlined_method_with_reordered_keywords
-    # The caller passes the keywords in (b, a) order while the callee declares
-    # them in (a, b) order, so the inliner must reorder per the keyword names
-    # rather than positional order. The subtraction is asymmetric so a wrong
-    # mapping would produce -9 instead of 9.
-    assert_runs '9', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def callee(a:, b:) = a - b
-      def test = callee(b: 1, a: 10)
-
-      test
-    RUBY
-  end
-
-  def test_inlined_method_with_keyword_default_using_prior_param
-    # When the caller omits an optional keyword whose default expression
-    # references another parameter, the default must execute in callee scope
-    # with the freshly-bound parameter visible.
-    assert_runs '202', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def callee(x, y: x * 100) = x + y
-      def test(n) = callee(n)
-
-      test(2)
-    RUBY
-  end
-
-  def test_inlined_method_with_rescue_caught_in_callee
-    # The callee's begin/rescue catches an exception raised inside the same
-    # callee. The runtime exception walker must find the rescue clause via the
-    # inlined callee's CFP (whose iseq pointer is the callee iseq), then resume
-    # the callee at the catch entry's continuation PC. The rescue branch returns
-    # 42 while the happy branch would return 0, so a wrong control-flow path
-    # would change the result.
-    assert_runs '42', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def callee(x)
-        begin
-          raise "boom" if x.negative?
-          0
-        rescue
-          42
-        end
-      end
-      def test(n) = callee(n)
-
-      test(-1)
-    RUBY
-  end
-
-  def test_inlined_method_with_rescue_caught_in_caller
-    # The callee re-raises (no rescue) and the caller catches the exception.
-    # This exercises the unwind path that pops the inlined callee's CFP and
-    # finds the catch entry on the *caller's* iseq, with the caller CFP's PC
-    # last stamped within the begin block by gen_save_pc_for_gc.
-    assert_runs '99', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def callee(x)
-        raise "boom" if x.negative?
-        0
-      end
-      def test(n)
-        begin
-          callee(n)
-        rescue
-          99
-        end
-      end
-
-      test(-1)
-    RUBY
-  end
-
-  def test_inlined_method_with_ensure_runs_on_propagation
-    # The callee has begin/ensure (no rescue). The ensure must execute its
-    # side effect before the exception propagates to the caller's rescue.
-    # Comparing the side-effect log to the rescue value lets us tell whether
-    # the ensure block actually ran on the propagation path.
-    assert_runs '"ensured: caught"', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      $log = []
-      def callee(x)
-        begin
-          raise "boom" if x.negative?
-          $log << "no_raise"
-        ensure
-          $log << "ensured"
-        end
-      end
-      def test(n)
-        begin
-          callee(n)
-          "no_rescue"
-        rescue
-          "caught"
-        end
-      end
-
-      result = test(-1)
-      "\#{$log.first}: \#{result}"
-    RUBY
-  end
-
-  def test_inlined_method_with_retry_resumes_begin_block
-    # The retry path (CATCH_TYPE_RETRY) sets cfp->pc to the begin block's
-    # start in the callee iseq. Without retry support on inlined frames this
-    # would either loop forever, segfault, or return the wrong count. The
-    # counter terminates the loop on the second pass so the test is bounded.
-    assert_runs '2', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def callee(counter)
-        begin
-          counter[0] += 1
-          raise "boom" if counter[0] < 2
-          counter[0]
-        rescue
-          retry
-        end
-      end
-      def test(c) = callee(c)
-
-      test([0])
-    RUBY
-  end
-
-  def test_inlined_method_with_super_call
-    assert_runs '"hi!"', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      class Parent
-        def greet = 'hi'
-      end
-
-      class Child < Parent
-        def greet = super + '!'
-      end
-
-      def test(c) = c.greet
-
-      child = Child.new
-      test(child)
-      test(child)
-    RUBY
-  end
-
-  def test_inlined_method_with_block_break_across_inlined_boundary
-    # When the inlined callee passes a literal block to a yielding method,
-    # the parent callee iseq carries CATCH_TYPE_BREAK entries covering the
-    # send opcode. A `break` from the block raises TAG_BREAK and unwinds to
-    # the inlined callee's CFP, where the catch entry must match. Asymmetric
-    # values (the break path yields 7, the run-to-completion path would
-    # yield the last element 99) discriminate between the two paths.
-    assert_runs '7', <<~RUBY, extra_args: ['--zjit-inline-threshold=30']
-      def callee(arr)
-        arr.each do |x|
-          break 7 if x > 5
-        end
-      end
-      def test(a) = callee(a)
-
-      test([1, 6, 99])
-    RUBY
-  end
-
   def test_exit_tracing
     # Smoke test: --zjit-trace-exits writes a Fuchsia trace (.fxt) file to /tmp
     assert_compiles('true', <<~RUBY, extra_args: ['--zjit-trace-exits'])
